# GGRC-5178 Fix memory leak for comments form