@tor4 2018-07-04
@YuriChupyrkin Please describe performance improvement for point 2 in more detail.